requirements.txt
# Flask Web Framework
flask==3.0.0
flask-sqlalchemy==3.1.1
flask-login==0.6.3
flask-wtf==1.2.1
wtforms==3.1.2

# Database
sqlalchemy==2.0.23

# PyTorch CPU only
torch==2.1.0
torchvision==0.16.0
torchaudio==2.1.0
--index-url https://download.pytorch.org/whl/cpu

# Transformers for DistilBERT
transformers==4.35.0

# Computer Vision
opencv-python==4.8.1.78
pillow==10.1.0

# OCR
pytesseract==0.3.10

# Image Processing
scikit-image==0.22.0
numpy==1.26.2

# PDF Generation
reportlab==4.0.7

# Utilities
werkzeug==3.0.1
python-dotenv==1.0.0
email-validator==2.1.0
itsdangerous==2.1.2
